Split generators.ts into multiple files matching hegel-rust layout
src/generators/
  core.ts         — Generator, BasicGenerator, combinators (map/filter/flatMap)
  numeric.ts      — integers, bigIntegers, floats, booleans
  strings.ts      — text, characters, binary, fromRegex, format generators
  collections.ts  — arrays/lists, sets, maps/dicts
  combinators.ts  — just, sampledFrom, oneOf, optional
  tuples.ts       — tuples, tuples3, tuples4
  compose.ts      — composite, record
  index.ts        — barrel re-exports
