+++ conformance/test_floats.ts
@@ -9,7 +9,7 @@
  */
 
 import { getTestCases, writeMetrics } from "../src/conformance.js";
-import { floats } from "../src/generators.js";
+import { floats } from "../src/generators/index.js";
 import { hegel } from "../src/runner.js";
 
 const params: Record<string, unknown> = process.argv[2] ? JSON.parse(process.argv[2]) : {};
