+++ conformance/test_hashmaps.ts
@@ -12,7 +12,7 @@
  */
 
 import { getTestCases, writeMetrics } from "../src/conformance.js";
-import { maps, integers, text } from "../src/generators.js";
+import { maps, integers, text } from "../src/generators/index.js";
 import { hegel } from "../src/runner.js";
 
 const params: Record<string, unknown> = process.argv[2] ? JSON.parse(process.argv[2]) : {};
