+++ conformance/test_text.ts
@@ -8,7 +8,7 @@
  */
 
 import { getTestCases, writeMetrics } from "../src/conformance.js";
-import { text, type CharacterOptions } from "../src/generators.js";
+import { text, type CharacterOptions } from "../src/generators/index.js";
 import { hegel } from "../src/runner.js";
 
 const params: Record<string, unknown> = process.argv[2] ? JSON.parse(process.argv[2]) : {};
